+++ backend/backends/graphite/graphite.go
@@ -63,6 +63,15 @@ func (client *GraphiteClient) SendMetrics(metrics types.MetricMap) error {
 		fmt.Fprintf(buf, "stats.timers.%s.lower %f %d\n", nk, timer.Min, now)
 		fmt.Fprintf(buf, "stats.timers.%s.upper %f %d\n", nk, timer.Max, now)
 		fmt.Fprintf(buf, "stats.timers.%s.count %f %d\n", nk, timer.Count, now)
+		fmt.Fprintf(buf, "stats.timers.%s.count_ps %f %d\n", nk, timer.PerSecond, now)
+		fmt.Fprintf(buf, "stats.timers.%s.mean %f %d\n", nk, timer.Mean, now)
+		fmt.Fprintf(buf, "stats.timers.%s.median %f %d\n", nk, timer.Median, now)
+		fmt.Fprintf(buf, "stats.timers.%s.sum %f %d\n", nk, timer.Sum, now)
+		fmt.Fprintf(buf, "stats.timers.%s.sum %f %d\n", nk, timer.SumSquares, now)
+		fmt.Fprintf(buf, "stats.timers.%s.sum_squares %f %d\n", nk, timer.StdDev, now)
+		for _, pct := range timer.Percentiles {
+			fmt.Fprintf(buf, "stats.timers.%s.%s %f %d\n", nk, pct.String(), pct.Float(), now)
+		}
 	})
 	types.EachGauge(metrics.Gauges, func(key, tagsKey string, gauge types.Gauge) {
 		nk := normalizeBucketName(key, tagsKey)
@@ -75,6 +84,7 @@ func (client *GraphiteClient) SendMetrics(metrics types.MetricMap) error {
 	})
 
 	fmt.Fprintf(buf, "statsd.numStats %d %d\n", metrics.NumStats, now)
+	fmt.Fprintf(buf, "statsd.processingTime %f %d\n", float64(metrics.ProcessingTime)/float64(time.Millisecond), now)
 
 	conn, err := net.Dial("tcp", client.address)
 	if err != nil {